feat(backup): Improve feature code structure
